Fix migrations errors

+++ components/shared/app/models/shared/assignment.rb
@@ -28,6 +28,8 @@ class Assignment < ActiveRecord::Base
   end
 
   def add_assignments_to_descendants
+    return true unless Shared::Assignment.column_names.include? 'admin_visible'
+
     if admin_visible
       delete_assignments_from_descendants
 
@@ -40,6 +42,8 @@ class Assignment < ActiveRecord::Base
   end
 
   def delete_assignments_from_descendants
+    return true unless Shared::Assignment.column_names.include? 'admin_visible'
+
     Shared::Assignment.where({ parent: id }).destroy_all if admin_visible
 
     true
